mix-demo-client-azstaticwebapps/api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-core==1.13.0
azure-functions==1.7.0
azure-search-documents==11.1.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
requests==2.25.1
requests-oauthlib==1.3.0
sendgrid==6.6.0
six==1.15.0
typing-extensions==3.10.0.0
urllib3==1.26.5
uuid


